% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsenm.R
\docType{package}
\name{ellipsenm-package}
\alias{ellipsenm}
\alias{ellipsenm-package}
\title{ellipsenm: An R package for ecological niche's characterization using ellipsoids}
\description{
A collection of tools to characterize ecological niches using ellipsoid envelopes. Functions to perform model calibration and selection, prepare model projections to distinct scenarios, assess niche overlap, and produce virtual ecological niches are included as part of this package. Other functions implemented here are useful to perform a series of common pre- and post-modeling analyses. In contrast to other tools for ecological niche and species distribution modeling, this package is the first in presenting a comprehensive set of tools that use ellipsoid envelopes as the theoretical representation of ecological niches. Details on the theoretical background and methods used here can be found in: Peterson et al. (1999) \doi{10.1126/science.285.5431.1265}, Soberon and Peterson (2005) \doi{10.17161/bi.v2i0.4}, Peterson (2011) \doi{10.1111/j.1365-2699.2010.02456.x}, Peterson et al. (2011) \url{https://bit.ly/2UpQ9Hd}
}
\details{
ellipsenm implements multiple tools to help users in using ellipsoids to model
ecological niches of species. Handy options for calibrating and selecting
models, producing models with replicates and projections, and assessing niche
overlap are included as part of this package. Other functions implemented here
are useful to perform a series of pre- and post-modeling analyses.
}
\section{Main functions in ellipsenm}{

\code{\link{ellipsoid_fit}}, \code{\link{predict}},
\code{\link{ellipsoid_calibration}}, \code{\link{prepare_sets}},
\code{\link{ellipsoid_model}}, \code{\link{overlap_object}},
\code{\link{ellipsoid_overlap}}, \code{\link{plot_overlap}}
}

\section{Other functions (important helpers)}{

\code{\link{data_subsample}}, \code{\link{ellipsoid_volume}},
\code{\link{report_format}}, \code{\link{mbased_mve}},
\code{\link{mmm_ellipsoid}}, \code{\link{model_projection}},
\code{\link{select_best}},
\code{\link{hypercube_boundaries}}, \code{\link{overlap_metrics}}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/marlonecobos/ellipsenm2}
  \item Report bugs at \url{https://github.com/marlonecobos/ellipsenm2/issues}
}

}
\author{
\strong{Maintainer}: Marlon E. Cobos \email{manubio13@gmail.com}

Authors:
\itemize{
  \item Luis Osorio-Olvera \email{luismurao@gmail.com }
  \item Jorge Soberon \email{jsoberon@ku.edu}
  \item A. Townsend Peterson \email{town@ku.edu}
  \item Vijay Barve \email{vijay.barve@gmail.com}
  \item Narayani Barve \email{narayani.ku@gmail.com}
}

}
